Fixed some win64 warnings

+++ src/prodbg/session/session.cpp
@@ -75,7 +75,7 @@ void Session_createDockingGrid(Session* session, int width, int height)
 
 bool Session_loadLayout(Session* session, const char* filename, int width, int height)
 {
-	UIDockingGrid* grid = UIDock_loadLayout(filename, width, height);
+	UIDockingGrid* grid = UIDock_loadLayout(filename, (float)width, (float)height);
 
 	if (!grid)
 		return false;
